Organisations can now have multiple locations

+++ app/Observers/Concerns/HasLocations.php
@@ -6,11 +6,13 @@ use App\Models\Creature;
 use App\Models\Location;
 use App\Models\MiscModel;
 
+use function PHPUnit\Framework\isEmpty;
+
 trait HasLocations
 {
     /**
      */
-    public function saveLocations(MiscModel|Creature $model)
+    public function saveLocations(MiscModel|Creature $model, array $locations = [])
     {
         /** @var Creature $model */
         $existing = $unique = $recreate = [];
@@ -28,11 +30,13 @@ trait HasLocations
         if (!empty($recreate)) {
             $model->locations()->attach($recreate);
         }
-
-        $locations = request()->get('locations', []);
+        if (!$locations) {
+            $locations = request()->get('locations', []);
+            $detach = true;
+        }
         $newLocations = [];
         foreach ($locations as $id) {
-            // Existing race, do nothing
+            // Existing location, do nothing
             if (!empty($existing[$id])) {
                 unset($existing[$id]);
                 continue;
@@ -51,7 +55,7 @@ trait HasLocations
         $model->locations()->attach($newLocations);
 
         // Detach the remaining
-        if (!empty($existing)) {
+        if (!empty($existing) && isset($detach)) {
             $model->locations()->detach($existing);
         }
     }
